Work on binary expressions
